refactor: create shareablity folders
