+++ src/lib/shareability/researcher/researchBriefPdf.jsx
