+++ src/lib/shareability/researcher/researchPanelAccent.js
