+++ src/pages/HeftiResearch.jsx
@@ -9,8 +9,8 @@ import { Heading } from '../components/ui/atom/heading';
 import { chartToRows } from '../components/ui/molecule/researchChart';
 import ResearchChatPanel from '../components/ui/organism/researchChatPanel';
 import ResearchChartsPanel from '../components/ui/organism/researchChartsPanel';
-import { getPanelAccent } from '../lib/shareability/researchPanelAccent';
-import { createResearchShareActions } from '../lib/shareability/researchShareActions';
+import { getPanelAccent } from '../lib/shareability/researcher/researchPanelAccent';
+import { createResearchShareActions } from '../lib/shareability/researcher/researchShareActions';
 import useResearchScroll from '../hooks/useResearchScroll';
 import useResearchContextCharts from '../hooks/useResearchContextCharts';
 import useResearchStream from '../hooks/useResearchStream';
